BRANDON: Los permisos ahora son validados para todos los modelos y vistas disponibles del sitio

+++ app/Http/Controllers/ClienteController.php
@@ -2,14 +2,15 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\User; // Asegúrate de que User es el modelo que utilizas
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ClienteController extends Controller
 {
     public function __construct()
     {
-        $this->middleware('auth', ['except' => ['index', 'show']]);
+        $this->middleware('auth');
     }
 
     /**
@@ -17,6 +18,7 @@ class ClienteController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', User::class);
         $clientes = User::all();  // Obtiene todos los usuarios
         return view("listado-clientes", compact("clientes"));  // Asegúrate de que la vista exista
     }
@@ -26,6 +28,7 @@ class ClienteController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', User::class);
         return view("crear-cliente");  // Cambia esto si es necesario
     }
 
@@ -34,6 +37,7 @@ class ClienteController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', User::class);
         $request->validate([
             'name' => ['required', 'max:255'], // Campo de nombre
             'apellido' => ['required', 'max:255'], // Campo de apellido
@@ -61,6 +65,7 @@ class ClienteController extends Controller
      */
     public function show(User $cliente)
     {
+        $this->authorize('view', $cliente);
         return view('show-cliente', compact('cliente'));  
     }
 
@@ -69,6 +74,7 @@ class ClienteController extends Controller
      */
     public function edit(User $cliente)
     {
+        $this->authorize('update', User::class);
         return view("edit-cliente", compact("cliente"));  
     }
 
@@ -77,6 +83,7 @@ class ClienteController extends Controller
      */
     public function update(Request $request, User $cliente)
     {
+        $this->authorize('update', User::class);
         $request->validate([
             'name' => ['required', 'max:255'],
             'apellido' => ['required', 'max:255'],
@@ -96,6 +103,7 @@ class ClienteController extends Controller
      */
     public function destroy(User $cliente)
     {
+        $this->authorize('update', User::class);
         $cliente->delete();
         return redirect()->route('cliente.index');  
     }
